removed projets subpage
