+++ components/project/projects.module.scss
@@ -1,0 +1,4 @@
+.projects {
+  display: flex;
+  justify-content: center;
+}
