+++ pages/index.js
@@ -1,7 +1,7 @@
 import styles from '../styles/Home.module.scss';
 import { Fragment } from 'react';
 import AboutMe from './about-me';
-import Projects from './projects';
+import Projects from '../components/project/projects';
 import ContactMe from './contact-me';
 import { useRef, useEffect } from 'react';
 import { useNavContext } from '../context/state';
