+++ pages/projects/index.js
@@ -1,45 +1,20 @@
-import Project from '../../components/project/project';
-import styles from './index.module.scss';
-import { forwardRef } from 'react';
+// import ProjectItem from '../../components/project/project-item';
+// import styles from './index.module.scss';
+// import { forwardRef } from 'react';
 
-const Projects = forwardRef((props, ref) => {
-  return (
-    <div className={styles.container}>
-      <h1 ref={ref}>My Projects</h1>
-      <div className={styles.projects}>
-        <Project
-          title={"GymAlly"}
-          img={"/gymally_screenshot.png"}
-          url={"http://www.gymally.fitness/"}
-          description={"Here is the description"}
-          skills={["Ruby", "JS"]}
-        />
-        <Project
-          title={"Help Local"}
-          img={"/helplocal_screenshot.png"}
-          url={"https://new-in-germany.vercel.app/"}
-          description={"Here is the description"}
-          skills={["JS", "Next.js", "MongoDB"]}
-        />
-      </div>
-    </div>
-  )
-});
-Projects.displayName = "Projects";
-
-// function Projects() {
+// const Projects = forwardRef((props, ref) => {
 //   return (
 //     <div className={styles.container}>
-//       <h1>My Projects</h1>
+//       <h1 ref={ref}>My Projects</h1>
 //       <div className={styles.projects}>
-//         <Project
+//         <ProjectItem
 //           title={"GymAlly"}
 //           img={"/gymally_screenshot.png"}
 //           url={"http://www.gymally.fitness/"}
 //           description={"Here is the description"}
 //           skills={["Ruby", "JS"]}
 //         />
-//         <Project
+//         <ProjectItem
 //           title={"Help Local"}
 //           img={"/helplocal_screenshot.png"}
 //           url={"https://new-in-germany.vercel.app/"}
@@ -49,6 +24,7 @@ Projects.displayName = "Projects";
 //       </div>
 //     </div>
 //   )
-// }
+// });
+// Projects.displayName = "Projects";
 
-export default Projects;
+// export default Projects;
