+++ pages/projects/index.module.scss
@@ -1,4 +1,4 @@
-.projects {
-  display: flex;
-  justify-content: center;
-}
+// .projects {
+//   display: flex;
+//   justify-content: center;
+// }
